move swap freq-ch function to preprocess
