+++ echopype/preprocess/__init__.py
@@ -1,3 +1,3 @@
-from .api import compute_MVBS, compute_MVBS_index_binning, remove_noise
+from .api import compute_MVBS, compute_MVBS_index_binning, remove_noise, swap_dims_channel_frequency
 
-__all__ = ["compute_MVBS", "compute_MVBS_index_binning", "remove_noise"]
+__all__ = ["compute_MVBS", "compute_MVBS_index_binning", "remove_noise", "swap_dims_channel_frequency"]
