--- echopype/utils/__init__.py
@@ -1,5 +1,0 @@
-from .common import swap_dims_channel_frequency
-
-__all__ = [
-    "swap_dims_channel_frequency",
-]
